// other/src/shared/types/interfaces.ts
export interface FilterItem {
  id: string;
  name: string;
  parentId?: string;
  type: AttributeTypeEnum;
  value: number;
}